+++ holonote/tests/test_annotators_basic.py
@@ -8,15 +8,12 @@
 # .snapshot() and .revert_to_snapshot()
 
 import uuid
-import unittest
-import pytest
+
 import numpy as np
 import pandas as pd
+import pytest
 
-import holoviews as hv
-from holonote.annotate import AnnotationTable
 from holonote.annotate import Annotator
-from holonote.annotate import SQLiteDB, UUIDHexStringKey
 
 
 @pytest.fixture()
@@ -159,7 +156,7 @@ class TestBasicRange1DAnnotator:
         descriptions = ['Annotation %d' % d for d in [1,2,3]]
         annotation_id1 = [uuid.uuid4().hex[:8] for d in [1,2,3]]
         mismatched = [uuid.uuid4().hex[:8] for d in [1,2]]
-        annotation_id2 = mismatched + [annotation_id1[2]]
+        annotation_id2 = [*mismatched, annotation_id1[2]]
 
         data1 = pd.DataFrame({'uuid':annotation_id1, 'start':starts,
                               'end':ends, 'description':descriptions}).set_index('uuid')
@@ -277,7 +274,7 @@ class TestBasicRange2DAnnotator:
         descriptions = ['Annotation %d' % d for d in [1,2,3]]
         annotation_id1 = [uuid.uuid4().hex[:8] for d in [1,2,3]]
         mismatched = [uuid.uuid4().hex[:8] for d in [1,2]]
-        annotation_id2 = mismatched + [annotation_id1[2]]
+        annotation_id2 = [*mismatched, annotation_id1[2]]
 
         data1 = pd.DataFrame({'uuid':annotation_id1, 'xstart':xstarts, 'xend':xends,
                              'ystart':ystarts, 'yend':yends,
@@ -385,7 +382,7 @@ class TestBasicPoint1DAnnotator:
         descriptions = ['Annotation %d' % d for d in [1,2,3]]
         annotation_id1 = [uuid.uuid4().hex[:8] for d in [1,2,3]]
         mismatched = [uuid.uuid4().hex[:8] for d in [1,2]]
-        annotation_id2 = mismatched + [annotation_id1[2]]
+        annotation_id2 = [*mismatched, annotation_id1[2]]
 
         data1 = pd.DataFrame({'uuid':annotation_id1, 'timestamps':timestamps,
                               'description':descriptions}).set_index('uuid')
@@ -393,9 +390,8 @@ class TestBasicPoint1DAnnotator:
                               'description':descriptions}).set_index('uuid')
 
         annotator_point1d.define_fields(data1[['description']])
-        with pytest.raises(KeyError, match=str(mismatched)) as cm:
+        with pytest.raises(KeyError, match=str(mismatched)):
             annotator_point1d.define_points(data2['timestamps'])
-
 
 
 class TestBasicPoint2DAnnotator:
@@ -493,7 +489,7 @@ class TestBasicPoint2DAnnotator:
         descriptions = ['Annotation %d' % d for d in [1,2,3]]
         annotation_id1 = [uuid.uuid4().hex[:8] for d in [1,2,3]]
         mismatched = [uuid.uuid4().hex[:8] for d in [1,2]]
-        annotation_id2 = mismatched + [annotation_id1[2]]
+        annotation_id2 = [*mismatched, annotation_id1[2]]
 
         data1 = pd.DataFrame({'uuid':annotation_id1, 'xs':xs, 'ys':ys,
                               'description':descriptions}).set_index('uuid')
@@ -503,132 +499,3 @@ class TestBasicPoint2DAnnotator:
         annotator_point2d.define_fields(data1[['description']])
         with pytest.raises(KeyError, match=str(mismatched)):
             annotator_point2d.define_points(data2['xs'], data2['ys'])
-
-
-
-class TestMultipleRegion1DAnnotator(unittest.TestCase):
-
-    def setUp(self):
-        assert Annotator.connector_class is SQLiteDB, 'Expecting default SQLite connector'
-        Annotator.connector_class.filename = ':memory:'
-        Annotator.connector_class.primary_key = UUIDHexStringKey()
-        self.annotator = Annotator({'TIME': np.datetime64}, fields=['description'],
-                                   region_types=['Point', 'Range'])
-
-    def tearDown(self):
-        self.annotator.connector.cursor.close()
-        self.annotator.connector.con.close()
-        del self.annotator
-
-    def test_point_range_commit_insertion(self):
-        descriptions = ['A point insertion', 'A range insertion']
-        timestamp = np.datetime64('2022-06-06')
-        self.annotator.set_point(timestamp)
-        self.annotator.add_annotation(description=descriptions[0])
-
-        start, end = np.datetime64('2022-06-06'), np.datetime64('2022-06-08')
-        self.annotator.set_range(start, end)
-        self.annotator.add_annotation(description=descriptions[1])
-
-        self.annotator.commit()
-
-        # FIXME! Index order is inverted?
-        df = pd.DataFrame({'uuid': pd.Series(self.annotator.df.index[::-1], dtype=object),
-                           'point_TIME':[timestamp, pd.NaT],
-                           'start_TIME':[pd.NaT, start],
-                           'end_TIME': [pd.NaT, end],
-                           'description':descriptions}
-                           ).set_index('uuid')
-
-        sql_df = self.annotator.connector.load_dataframe()
-        pd.testing.assert_frame_equal(sql_df, df)
-
-
-class TestMultiplePlotAnnotator(unittest.TestCase):
-
-    def setUp(self):
-        assert Annotator.connector_class is SQLiteDB, 'Expecting default SQLite connector'
-        Annotator.connector_class.filename = ':memory:'
-        Annotator.connector_class.primary_key = UUIDHexStringKey()
-
-        self.connector = SQLiteDB()
-        xvals, yvals  = np.linspace(-4, 0, 202), np.linspace(4, 0, 202)
-        xs, ys = np.meshgrid(xvals, yvals)
-        image = hv.Image(np.sin(ys*xs), kdims=['A', 'B'])
-        self.image_annotator = Annotator(image, connector=self.connector,
-                                         fields=['description'], region_types=['Range'])
-
-        curve = hv.Curve((np.arange('2005-02', '2005-03', dtype='datetime64[D]'), range(28)), kdims=['TIME'])
-        self.curve_annotator = Annotator(curve, connector=self.connector,
-                                         fields=['description'], region_types=['Range'])
-
-    def test_element_kdim_dtypes(self):
-        assert self.image_annotator.kdim_dtypes == {'A': np.float64, 'B': np.float64}
-        assert self.curve_annotator.kdim_dtypes == {'TIME': np.datetime64}
-
-    def test_multiplot_add_annotation(self):
-        self.image_annotator.set_range(-0.25, 0.25, -0.1, 0.1)
-        self.curve_annotator.set_range(np.datetime64('2005-02-13'), np.datetime64('2005-02-16'))
-        self.connector.add_annotation(description='Multi-plot annotation')
-
-
-
-    def tearDown(self):
-        self.connector.cursor.close()
-        self.connector.con.close()
-        del self.image_annotator
-
-
-class TestAnnotatorMultipleStringFields(unittest.TestCase):
-
-    def setUp(self):
-        assert Annotator.connector_class is SQLiteDB, 'Expecting default SQLite connector'
-        Annotator.connector_class.filename = ':memory:'
-        Annotator.connector_class.primary_key = UUIDHexStringKey()
-        self.annotator = Annotator({'TIME': np.datetime64}, fields=['field1', 'field2'])
-
-
-    def test_insertion_values(self):
-        start, end = np.datetime64('2022-06-06'), np.datetime64('2022-06-08')
-        self.annotator.set_range(start, end)
-        self.annotator.add_annotation(field1='A test field', field2='Another test field')
-        commits = self.annotator.annotation_table.commits()
-        kwargs = commits[0]['kwargs']
-        assert len(commits)==1, 'Only one insertion commit made'
-        assert 'uuid' in kwargs.keys(), 'Expected uuid primary key in kwargs'
-        kwargs.pop('uuid')
-        assert kwargs == dict(field1='A test field', field2='Another test field', start_TIME=start, end_TIME=end)
-
-
-    def test_commit_insertion(self):
-        start, end  = np.datetime64('2022-06-06'), np.datetime64('2022-06-08')
-        field1 = 'A test field'
-        field2 = 'Another test field'
-        self.annotator.set_range(start, end)
-        self.annotator.add_annotation(field1=field1, field2=field2)
-        self.annotator.commit()
-
-        df = pd.DataFrame({'uuid': pd.Series(self.annotator.df.index[0], dtype=object),
-                           'start_TIME':[start],
-                           'end_TIME':[end],
-                           'field1':[field1],
-                           'field2':[field2]}
-                           ).set_index('uuid')
-
-        sql_df = self.annotator.connector.load_dataframe()
-        pd.testing.assert_frame_equal(sql_df, df)
-
-
-    def test_commit_update(self):
-        start1, end1  = np.datetime64('2022-06-06'), np.datetime64('2022-06-08')
-        start2, end2  = np.datetime64('2023-06-06'), np.datetime64('2023-06-08')
-        start3, end3  = np.datetime64('2024-06-06'), np.datetime64('2024-06-08')
-        self.annotator.set_range(start1, end1)
-        self.annotator.add_annotation(field1='Field 1.1', field2='Field 1.2')
-        self.annotator.set_range(start2, end2)
-        self.annotator.add_annotation(field1='Field 2.1', field2='Field 2.2')
-        self.annotator.commit()
-        self.annotator.update_annotation_fields(self.annotator.df.index[0], field1='NEW Field 1.1')
-        self.annotator.commit()
-        sql_df = self.annotator.connector.load_dataframe()
-        assert set(sql_df['field1']) == set(['NEW Field 1.1', 'Field 2.1'])
